requirements.txt
asgiref==3.5.0
certifi==2021.10.8
Django==3.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
pytz==2021.3
sqlparse==0.4.2
wincertstore==0.2
